+++ src/main/scala/me/welcomer/framework/eci/EciResolver.scala
@@ -10,7 +10,7 @@
  *  distributed under the License is distributed on an "AS IS" BASIS,
  *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  *  See the License for the specific language governing permissions and
- *  limitations under the License. 
+ *  limitations under the License.
  */
 package me.welcomer.framework.eci
 
@@ -82,7 +82,7 @@ private[framework] class EciResolver(
       resolvedEci map { optionEci =>
         val resolveEvent: EciResolverMsg = optionEci match {
           case Some(picoUUID) => ECIResolvedToPicoUUID(eci, picoUUID)
-          case None => InvalidECI(eci)
+          case None           => InvalidECI(eci)
         }
 
         //        eciResolverBus.publish(resolveEvent)