+++ src/main/scala/me/welcomer/framework/eventedgateway/EventedGateway.scala
@@ -10,7 +10,7 @@
  *  distributed under the License is distributed on an "AS IS" BASIS,
  *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  *  See the License for the specific language governing permissions and
- *  limitations under the License. 
+ *  limitations under the License.
  */
 package me.welcomer.framework.eventedgateway
 
@@ -81,7 +81,7 @@ private[framework] class EventedGateway(
     case EventedEntityResolver.EntityResolveFailure(evented, error) => {
       evented.replyTo map { replyTo =>
         replyTo ! EventedFailure(error)
-      } getOrElse { log.debug("{} ({})", error, evented.entityId) }
+      } getOrElse { log.error("{} ({})", error, evented.entityId) }
     }
   }